+++ resources/views/components/layouts/setup.blade.php
@@ -73,6 +73,10 @@
                 'url' : '/auth/login'
             },
             {
+                'name' : 'Passkeys Setup',
+                'url' : '/auth/setup/passkeys'
+            },
+            {
                 'name' : 'Register',
                 'url' : '/auth/register'
             },